pyproject.toml
[project]
name = "resume-analizer"
version = "0.1.0"
description = "An innovative AI-powered tool crafted to evaluate resumes and provide tailored, actionable insights, empowering job seekers to refine their applications."
readme = "README.md"
requires-python = ">=3.13"
license = { text = "MIT" }
authors = [{ name = "Ryzhaiev Serhii", email = "rsp89.we@gmail.com" }]

dependencies = [
    "openai>=1.78.0",
    "pypdf2>=3.0.1",
    "python-docx>=1.1.2",
    "python-dotenv>=1.1.0",
    "streamlit>=1.45.0",
]
